requirements.txt
# ╔══════════════════════════════════════════════════════════════╗
# ║      AI TRADING PLATFORM — PRODUCTION REQUIREMENTS          ║
# ║      Hedge-Fund Grade | Pinned Versions | Zero Conflicts     ║
# ╚══════════════════════════════════════════════════════════════╝

# ── Core Runtime ────────────────────────────────────────────────
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# ── Telegram ────────────────────────────────────────────────────
python-telegram-bot==20.7

# ── Exchange Connectivity ────────────────────────────────────────
ccxt==4.2.0  # Use ccxt[pro] for WebSocket: pip install ccxt[pro]

# ── Data Science Core ───────────────────────────────────────────
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# ── Technical Analysis ──────────────────────────────────────────
pandas-ta==0.3.14b0

# ── ML / AI ─────────────────────────────────────────────────────
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.2.0
optuna==3.5.0
joblib==1.3.2

# ── Deep Learning (optional) ────────────────────────────────────
# torch==2.1.2  # Uncomment for LSTM model (~700MB)

# ── Backtesting ─────────────────────────────────────────────────
backtrader==1.9.78.123

# ── ML Experiment Tracking ──────────────────────────────────────
mlflow==2.9.2

# ── Database ────────────────────────────────────────────────────
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
# NOTE: aioredis REMOVED — merged into redis>=4.2, causes import conflicts

# ── Web / API ───────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# ── Web Dashboard (Flask) ────────────────────────────────────────
Flask>=3.0.0
Jinja2>=3.1.0
itsdangerous>=2.1.0

# ── Auth ────────────────────────────────────────────────────────
PyJWT>=2.7.0
requests>=2.31.0

# ── Monitoring & Observability ───────────────────────────────────
prometheus-client==0.19.0
structlog==24.1.0

# ── Async Runtime ───────────────────────────────────────────────
aiohttp==3.9.1
asyncio-throttle==1.0.2

# ── Scheduling ──────────────────────────────────────────────────
APScheduler==3.10.4

# ── Reporting ───────────────────────────────────────────────────
matplotlib==3.8.2
seaborn==0.13.1
reportlab==4.0.8

# ── Data Serialization ──────────────────────────────────────────
pyarrow==14.0.2

# ── Security ────────────────────────────────────────────────────
cryptography>=41.0.0

# ── Testing ─────────────────────────────────────────────────────
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
httpx==0.26.0

# ── Forex / MT5 ─────────────────────────────────────────────────
# MetaTrader5==5.0.45  # Windows only
